copilot + cleanup
