+++ addons/h1/model.cfg
@@ -462,10 +462,10 @@ class CfgModels
 				selection="rucicka_07a";
 				axis="rucicka_07a_osa";
 			};			
-			class tgp_turet
+			class tgp_turret
 			{
 				type="rotationY";
-				source="ObsTurret";
+				source="tgp_turret";
 				selection="tgp_turret";
 				axis="tgp_dir_axis";
 				animPeriod=0;
@@ -474,10 +474,10 @@ class CfgModels
 				angle0="rad -360";
 				angle1="rad +360";
 			};                        
-			class tgp_ball: tgp_turet
+			class tgp_ball: tgp_turret
 			{
 				type="rotationX";
-                source="ObsGun";
+        source="tgp_ball";
 				selection="tgp_ball";
 				axis="tgp_elv_axis";
 			};
